+++ CHANGELOG.md
@@ -5,7 +5,17 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.7.4 - Unreleased
+## Version 0.7.5 - Unreleased
+
+### Changed
+* Added `imgs_with_annots` to extended stats.
+* Background param now works as an argument to `demo` for vidshapes
+
+### Fixed
+* Added missing `name_to_video` index in sql dataset
+
+
+## Version 0.7.4 - Released 2023-11-12
 
 ### Added